moved utils.js to utils
